TreeGen Testing Continues
Last bug reported was not so much a bug, but lack of clarity of terms. Does segnum mean actual chunks of segments or segment divisions. It refers to the latter now but could change it later as the former makes more sense.
Added testing tools to figure out what's going on with the TreeGen. These tools showed me that the problem lies within the treeMesh.verts population procedure (right now many of the verts locations are unassigned so that whole thing collapses to the origin).
These tools took a bit to put together but they are robust and able to be reused in future testing of procedurally created gameobject hierarchies.